'almost_done'

+++ lib/admin/ProgressTimeLine/ProgressPage.dart
@@ -21,48 +21,42 @@ class _ProgressPageState extends State<ProgressPage> {
   double percent2 = 80;
   final databaseReference = FirebaseDatabase.instance.reference();
   List allProjects = List();
-  int noOfProjects=0;
-  int completed=0;
-  int notStarted=0;
-  int ongoing=0;
-  double completedPercent=0.0;
-  double notStartedPercent=0.0;
-  double ongoingPercent=0.0;
+  int noOfProjects = 0;
+  int completed = 0;
+  int notStarted = 0;
+  int ongoing = 0;
+  double completedPercent = 0.0;
+  double notStartedPercent = 0.0;
+  double ongoingPercent = 0.0;
 
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
 
-    databaseReference.child("projects").once().then((DataSnapshot dataSnapshot) {
+    databaseReference
+        .child("projects")
+        .once()
+        .then((DataSnapshot dataSnapshot) {
       Map data = dataSnapshot.value;
       allProjects = [];
-      allProjects= data.values.toList();
+      allProjects = data.values.toList();
 
-      noOfProjects= allProjects.length;
-      for(int i=0; i<allProjects.length;i++)
-        {
-          var progPercent=double.parse(allProjects[i]["progressPercent"]);
-          if(progPercent>0 && progPercent<100)
-            ongoing++;
-          if(progPercent<=0)
-            notStarted++;
-          if(progPercent>=100)
-            completed++;
-        }
+      noOfProjects = allProjects.length;
+      for (int i = 0; i < allProjects.length; i++) {
+        var progPercent = double.parse(allProjects[i]["progressPercent"]);
+        if (progPercent > 0 && progPercent < 100) ongoing++;
+        if (progPercent <= 0) notStarted++;
+        if (progPercent >= 100) completed++;
+      }
 
       setState(() {
-        ongoingPercent= (ongoing/noOfProjects)*100;
-        notStartedPercent= (notStarted/noOfProjects)*100;
-        completedPercent= (completed/noOfProjects)*100;
+        ongoingPercent = (ongoing / noOfProjects) * 100;
+        notStartedPercent = (notStarted / noOfProjects) * 100;
+        completedPercent = (completed / noOfProjects) * 100;
 
         debugPrint(ongoingPercent.toString());
       });
-
-
     });
-
-
   }
 
   @override
@@ -85,7 +79,7 @@ class _ProgressPageState extends State<ProgressPage> {
                         animation: true,
                         percent: completedPercent / 100,
                         center: new Text(
-                          completedPercent.toString() + "%",
+                          completedPercent.toInt().toString() + "%",
                           style: new TextStyle(
                               fontWeight: FontWeight.bold, fontSize: 15.0),
                         ),
@@ -108,7 +102,7 @@ class _ProgressPageState extends State<ProgressPage> {
                         animation: true,
                         percent: ongoingPercent / 100,
                         center: new Text(
-                          ongoingPercent.toString() + "%",
+                          (ongoingPercent.toInt()).toString() + "%",
                           style: new TextStyle(
                               fontWeight: FontWeight.bold, fontSize: 15.0),
                         ),
@@ -129,9 +123,9 @@ class _ProgressPageState extends State<ProgressPage> {
                         radius: 60.0,
                         lineWidth: 6.0,
                         animation: true,
-                        percent: notStartedPercent/ 100,
+                        percent: notStartedPercent / 100,
                         center: new Text(
-                          notStartedPercent.toString() + "%",
+                          notStartedPercent.toInt().toString() + "%",
                           style: new TextStyle(
                               fontWeight: FontWeight.bold, fontSize: 15.0),
                         ),
